Refactor error handling by focussing on new submissions form
Errors pertaining to specific fields will display inline based on how  is structured.
If the error message is attached to :base, then only that message will show at the top of the form.
If the error message is associated with a specific attribute, it will display below the field within the form.

Some styling changes that make things good enough for now