WIP Move module with parent_vote(parent) back to lib.
I think I want this to be a view helper, rather than something in the
controller concerns - or it should be part of the model or in a querry
object.

I made some changes to the application helper in an effort to implement
the approach that Ryan Bates took in his postcast on presenters.